UUID system
This isn't a complete move.

On reason is that I removed auto-rejoin from the client as I plan to add it to the server-side later.

And the other reason is that currently the audience and speaker Maps have sockets as the keys, and I plan to change this because there should only be one state per UUID rather than per socket. Currently, the states between sockets are synced because they have the same socket ID, and this causes the state to be counted for each tab they have open.